+++ src/sim/mecafil.cc
@@ -142,7 +142,7 @@ void Mecafil::storeDirections()
      */
     
     // for the extremities, the direction of the nearby segment is used.
-    for ( size_t d = 0; d < DIM; ++d )
+    for ( unsigned d = 0; d < DIM; ++d )
     {
         rfDir[d]     = rfDiff[d];
         rfDir[d+end] = rfDiff[d+end-DIM];
@@ -675,7 +675,7 @@ void add_rigidityF(const size_t nbt, const real* X, const real R1, real* Y)
     real      * Z = Y + nbt + DIM;
     real const* E = X + nbt + DIM;
     #pragma ivdep
-    for ( int d = 0; d < DIM; ++d )
+    for ( unsigned d = 0; d < DIM; ++d )
     {
         Y[d    ] -= R1 * (X[d+DIM*2]+X[d]) - R2 * X[d+DIM];
         Y[d+DIM] -= R1 * (X[d+DIM]+X[d+DIM*3]) + R4 * (X[d+DIM]-X[d+DIM*2]) - R2 * X[d];
@@ -690,7 +690,7 @@ void add_rigidityF(const size_t nbt, const real* X, const real R1, real* Y)
  */
 void add_rigidity(size_t A, size_t B, size_t C, const real* X, const real R1, real* Y)
 {
-    for ( size_t d = 0; d < DIM; ++ d )
+    for ( unsigned d = 0;  d < DIM; ++ d )
     {
         real x = 2*X[B*DIM+d] - ( X[A*DIM+d] + X[C*DIM+d] );
         Y[A*DIM+d] += x * R1;
